Moved migration function to utils
